Move checkers to plugin directory
