+++ formver.compiler-plugin/locality/src/org/jetbrains/kotlin/formver/locality/plugin/ExpressionLocalityExtractor.kt
@@ -3,7 +3,7 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
-package org.jetbrains.kotlin.formver.locality.plugin.extension
+package org.jetbrains.kotlin.formver.locality.plugin
 
 import org.jetbrains.kotlin.fir.analysis.checkers.context.CheckerContext
 import org.jetbrains.kotlin.fir.declarations.FirProperty
