+++ formver.compiler-plugin/locality/src/org/jetbrains/kotlin/formver/locality/plugin/LocalityErrors.kt
@@ -8,7 +8,6 @@ package org.jetbrains.kotlin.formver.locality.plugin
 import com.intellij.psi.PsiElement
 import org.jetbrains.kotlin.diagnostics.KtDiagnosticsContainer
 import org.jetbrains.kotlin.diagnostics.error3
-import org.jetbrains.kotlin.formver.locality.plugin.extension.LocalityAttribute
 
 object LocalityErrors : KtDiagnosticsContainer() {
     val LOCALITY_VIOLATION by error3<PsiElement, String, LocalityAttribute?, LocalityAttribute?>()
