+++ formver.compiler-plugin/locality/src/org/jetbrains/kotlin/formver/locality/plugin/LocalityExtensionRegistrar.kt
@@ -6,8 +6,6 @@
 package org.jetbrains.kotlin.formver.locality.plugin
 
 import org.jetbrains.kotlin.fir.extensions.FirExtensionRegistrar
-import org.jetbrains.kotlin.formver.locality.plugin.extension.LocalityAdditionalCheckers
-import org.jetbrains.kotlin.formver.locality.plugin.extension.LocalityAttributeExtension
 
 object LocalityExtensionRegistrar : FirExtensionRegistrar() {
     override fun ExtensionRegistrarContext.configurePlugin() {
