+++ formver.compiler-plugin/locality/src/org/jetbrains/kotlin/formver/locality/plugin/LocalityReturnChecker.kt
@@ -3,28 +3,28 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
-package org.jetbrains.kotlin.formver.locality.plugin.extension
+package org.jetbrains.kotlin.formver.locality.plugin
 
 import org.jetbrains.kotlin.diagnostics.DiagnosticReporter
 import org.jetbrains.kotlin.diagnostics.reportOn
 import org.jetbrains.kotlin.fir.analysis.checkers.MppCheckerKind
 import org.jetbrains.kotlin.fir.analysis.checkers.context.CheckerContext
-import org.jetbrains.kotlin.fir.analysis.checkers.expression.FirThrowExpressionChecker
-import org.jetbrains.kotlin.fir.expressions.FirThrowExpression
+import org.jetbrains.kotlin.fir.analysis.checkers.expression.FirReturnExpressionChecker
+import org.jetbrains.kotlin.fir.expressions.FirReturnExpression
 import org.jetbrains.kotlin.formver.locality.plugin.LocalityErrors.LOCALITY_VIOLATION
 
-object LocalityThrowChecker : FirThrowExpressionChecker(MppCheckerKind.Common) {
+object LocalityReturnChecker : FirReturnExpressionChecker(MppCheckerKind.Common) {
     context(context: CheckerContext, reporter: DiagnosticReporter)
-    override fun check(expression: FirThrowExpression) {
+    override fun check(expression: FirReturnExpression) {
         val requiredLocality: LocalityAttribute? = null
-        val actualLocality = expression.exception.extractLocality()
+        val actualLocality = expression.result.extractLocality()
 
         if (requiredLocality.accepts(actualLocality)) return
 
         reporter.reportOn(
-            expression.exception.source,
+            expression.result.source,
             LOCALITY_VIOLATION,
-            "Throw",
+            "Return",
             requiredLocality,
             actualLocality
         )
