+++ formver.compiler-plugin/locality/src/org/jetbrains/kotlin/formver/locality/plugin/LocalityThrowChecker.kt
@@ -3,28 +3,28 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
-package org.jetbrains.kotlin.formver.locality.plugin.extension
+package org.jetbrains.kotlin.formver.locality.plugin
 
 import org.jetbrains.kotlin.diagnostics.DiagnosticReporter
 import org.jetbrains.kotlin.diagnostics.reportOn
 import org.jetbrains.kotlin.fir.analysis.checkers.MppCheckerKind
 import org.jetbrains.kotlin.fir.analysis.checkers.context.CheckerContext
-import org.jetbrains.kotlin.fir.analysis.checkers.expression.FirVariableAssignmentChecker
-import org.jetbrains.kotlin.fir.expressions.FirVariableAssignment
+import org.jetbrains.kotlin.fir.analysis.checkers.expression.FirThrowExpressionChecker
+import org.jetbrains.kotlin.fir.expressions.FirThrowExpression
 import org.jetbrains.kotlin.formver.locality.plugin.LocalityErrors.LOCALITY_VIOLATION
 
-object LocalityVariableAssignmentChecker : FirVariableAssignmentChecker(MppCheckerKind.Common) {
+object LocalityThrowChecker : FirThrowExpressionChecker(MppCheckerKind.Common) {
     context(context: CheckerContext, reporter: DiagnosticReporter)
-    override fun check(expression: FirVariableAssignment) {
-        val requiredLocality = expression.lValue.extractLocality()
-        val actualLocality = expression.rValue.extractLocality()
+    override fun check(expression: FirThrowExpression) {
+        val requiredLocality: LocalityAttribute? = null
+        val actualLocality = expression.exception.extractLocality()
 
         if (requiredLocality.accepts(actualLocality)) return
 
         reporter.reportOn(
-            expression.rValue.source,
+            expression.exception.source,
             LOCALITY_VIOLATION,
-            "Assignment",
+            "Throw",
             requiredLocality,
             actualLocality
         )
