+++ formver.compiler-plugin/locality/src/org/jetbrains/kotlin/formver/locality/plugin/LocalityVariableAssignmentChecker.kt
@@ -3,29 +3,28 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
-package org.jetbrains.kotlin.formver.locality.plugin.extension
+package org.jetbrains.kotlin.formver.locality.plugin
 
 import org.jetbrains.kotlin.diagnostics.DiagnosticReporter
 import org.jetbrains.kotlin.diagnostics.reportOn
 import org.jetbrains.kotlin.fir.analysis.checkers.MppCheckerKind
 import org.jetbrains.kotlin.fir.analysis.checkers.context.CheckerContext
-import org.jetbrains.kotlin.fir.analysis.checkers.declaration.FirValueParameterChecker
-import org.jetbrains.kotlin.fir.declarations.FirValueParameter
+import org.jetbrains.kotlin.fir.analysis.checkers.expression.FirVariableAssignmentChecker
+import org.jetbrains.kotlin.fir.expressions.FirVariableAssignment
 import org.jetbrains.kotlin.formver.locality.plugin.LocalityErrors.LOCALITY_VIOLATION
 
-object LocalityValueParameterChecker : FirValueParameterChecker(MppCheckerKind.Common) {
+object LocalityVariableAssignmentChecker : FirVariableAssignmentChecker(MppCheckerKind.Common) {
     context(context: CheckerContext, reporter: DiagnosticReporter)
-    override fun check(declaration: FirValueParameter) {
-        val defaultValue = declaration.defaultValue ?: return
-        val requiredLocality = declaration.extractRequiredLocality()
-        val actualLocality = defaultValue.extractLocality()
+    override fun check(expression: FirVariableAssignment) {
+        val requiredLocality = expression.lValue.extractLocality()
+        val actualLocality = expression.rValue.extractLocality()
 
         if (requiredLocality.accepts(actualLocality)) return
 
         reporter.reportOn(
-            defaultValue.source ?: declaration.source,
+            expression.rValue.source,
             LOCALITY_VIOLATION,
-            "Initializer",
+            "Assignment",
             requiredLocality,
             actualLocality
         )
